+++ src/main/java/com/teksystem/salestaxes/model/items/NonTaxableImportedItem.java
@@ -1,11 +1,13 @@
 package com.teksystem.salestaxes.model.items;
 
-import com.teksystem.salestaxes.utils.NegativeDecimalException;
 import com.teksystem.salestaxes.model.taxes.visitor.TaxVisitor;
+import com.teksystem.salestaxes.utils.NegativeDecimalException;
+
+import java.math.BigDecimal;
 
 public class NonTaxableImportedItem extends AbstractItem {
 
-    public NonTaxableImportedItem(final String name, final Double price) throws NegativeDecimalException {
+    public NonTaxableImportedItem(final String name, final BigDecimal price) throws NegativeDecimalException {
         super(name, price);
     }
