+++ src/main/java/com/teksystem/salestaxes/model/items/TaxableItem.java
@@ -1,11 +1,13 @@
 package com.teksystem.salestaxes.model.items;
 
-import com.teksystem.salestaxes.utils.NegativeDecimalException;
 import com.teksystem.salestaxes.model.taxes.visitor.TaxVisitor;
+import com.teksystem.salestaxes.utils.NegativeDecimalException;
+
+import java.math.BigDecimal;
 
 public class TaxableItem extends AbstractItem {
 
-    public TaxableItem(final String name, final Double price) throws NegativeDecimalException {
+    public TaxableItem(final String name, final BigDecimal price) throws NegativeDecimalException {
         super(name, price);
     }
